suppression et condition
Lors de la suppression d'un article, il sera aussi supprimé des lignes
dans la table basket

Condition pour l'affichage des images dans viewArt

+++ app/Controller/AjaxController.php
@@ -15,6 +15,7 @@ use \Model\ShippingModel;
 use \Model\FilterModel;
 use \Model\FiltrearticleModel;
 use \Model\SalesrevenueModel;
+use \Model\BasketModel;
 use \Model\FavoriteModel;
 use \W\Security\AuthentificationModel;
 use \PHPMailer;
@@ -69,9 +70,10 @@ class AjaxController extends Controller
 			if(is_numeric($_POST['id_item'])){
 				$favorite = new FavoriteModel();
 				$filtre = new FiltrearticleModel();
+				$basket = new BasketModel();
 				$itemModel  = new ItemModel();
 
-				if($favorite->deleteItem($_POST['id_item']) && $filtre->deleteByItem($_POST['id_item'])){
+				if($favorite->deleteItem($_POST['id_item']) && $filtre->deleteByItem($_POST['id_item']) && $basket->deleteItem($_POST['id_item'])){
 					$deleteItem = $itemModel->delete($_POST['id_item']);
 				}
